Revert "#96 feat : Add curation form (#102)"
This reverts commit f138b4831e38403d59c409f8a776dc328fa83b62.
